Rename launcher files
