+++ launcher/launcher.cpp
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include <list>
 
-#include "firmware.hpp"
+#include "launcher.hpp"
 
 #include "engine/api_private.hpp"
 #include "graphics/color.hpp"
